+++ tests/test_groq.py
@@ -3,6 +3,7 @@ import warnings
 warnings.simplefilter("ignore", DeprecationWarning)
 
 import unittest
+import threading
 
 from vectara_agentic.agent import Agent, AgentType
 from vectara_agentic.agent_config import AgentConfig
@@ -17,6 +18,9 @@ def mult(x: float, y: float) -> float:
     return x * y
 
 
+ARIZE_LOCK = threading.Lock()
+
+
 fc_config_groq = AgentConfig(
     agent_type=AgentType.FUNCTION_CALLING,
     main_llm_provider=ModelProvider.GROQ,
@@ -27,19 +31,20 @@ fc_config_groq = AgentConfig(
 class TestGROQ(unittest.TestCase):
 
     def test_multiturn(self):
-        tools = [ToolsFactory().create_tool(mult)]
-        topic = "AI topic"
-        instructions = "Always do as your father tells you, if your mother agrees!"
-        agent = Agent(
-            tools=tools,
-            topic=topic,
-            custom_instructions=instructions,
-        )
+        with ARIZE_LOCK:
+            tools = [ToolsFactory().create_tool(mult)]
+            topic = "AI topic"
+            instructions = "Always do as your father tells you, if your mother agrees!"
+            agent = Agent(
+                tools=tools,
+                topic=topic,
+                custom_instructions=instructions,
+            )
 
-        agent.chat("What is 5 times 10. Only give the answer, nothing else")
-        agent.chat("what is 3 times 7. Only give the answer, nothing else")
-        res = agent.chat("multiply the results of the last two questions. Output only the answer.")
-        self.assertEqual(res.response, "1050")
+            agent.chat("What is 5 times 10. Only give the answer, nothing else")
+            agent.chat("what is 3 times 7. Only give the answer, nothing else")
+            res = agent.chat("multiply the results of the last two questions. Output only the answer.")
+            self.assertEqual(res.response, "1050")
 
 
 if __name__ == "__main__":